backend/requirements.txt
# Web Framework and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM and Embeddings
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6

# Vector Database
chromadb==0.4.18

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Data Validation and Models
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging
loguru==0.7.2
